requirements.txt
streamlit
streamlit-chat
streamlit-extras
langchain
langchain-openai
python-dotenv
streamlit-lottie
